Unit test for compareSectionNumbers
